+++ P1/Assets/@Scripts/UI/SubItem/Character/UI_CharacterPanel.cs
@@ -84,7 +84,7 @@ public class UI_CharacterPanel : UI_Base
     // 전투력 표시 텍스트 함수 
     private void ShowTotalPowerText()
     {
-        GetTMPText((int)Texts.Text_TotalPower).text = Util.ConvertToTotalCurrency(Managers.Hero.PlayerHeroInfo.CurrentTotalPower);
+        GetTMPText((int)Texts.Text_TotalPower).text = Util.ConvertToTotalCurrency((long)Managers.Hero.PlayerHeroInfo.CurrentTotalPower);
     }
 
     // 장비 슬롯 클릭시 장비 팝업으로 이동하는 함수
